biblioteca/requirements.txt
asgiref==3.11.1
crispy-bootstrap4==2026.2
Django==6.0.3
django-crispy-forms==2.6
djangorestframework==3.17.1
sqlparse==0.5.5
tzdata==2025.3
